fix: adjust jest run script for ES6 modules
According to the documentation https://jestjs.io/docs/ecmascript-modules
in order to enable the support for ES6 modules when running jest one
needs to pass --experimental-vm-modules to node. While putting explicit
calls to node in the config for npm run-scripts is discouraged (see
https://docs.npmjs.com/cli/v8/commands/npm-run-script) so far I've not
been able to find a way to pass extra CLI options for node when running
via npm run so decided to go with the explicit command for now.