test: add test for reading from input file
The test reads from the file data/input.txt which is shipped with the
repository and contains the example input string. Mock output stream is
used to check that the contents of the file has been read.

TODO: learn how jest mocks can be used to replace fs calls to prevent
the need of reading from the real file on disk.