fix: make test portable
As I was able to discover when running on Linux fs.open() fails when
it's applied to a directory, while on Windows it's possible to "open"
the directory and the actual error happens on the write attempt to it.
I've changed the text of the error message which is checked in the test
to make it portable.